Check contact survey data

Checks that a survey fulfills all the requirements to work with the 'contact_matrix' function

as_contact_survey(
  x,
  id.column = "part_id",
  country.column = "country",
  year.column = "year"
)

Arguments

x

list containing

an element named 'participants', a data frame containing participant information

an element named 'contacts', a data frame containing contact information

(optionally) an element named 'reference, a list containing information information needed to reference the survey, in particular it can contain$a "title", "bibtype", "author", "doi", "publisher", "note", "year"

id.column

the column in both the participants and contacts data frames that links contacts to participants

country.column

the column in the participants data frame containing the country in which the participant was queried

year.column

the column in the participants data frame containing the year in which the participant was queried

Value

invisibly returns a character vector of the relevant columns

Examples

if (FALSE) { # \dontrun{
# not run because it requires an internet connection
current_surveys <- list_surveys()
polymod_url <- subset(current_surveys, grepl("POLYMOD", title))[["url"]]
# "https://doi.org/10.5281/zenodo.3874557"
polymod <- get_survey(polymod_url)
polymod
as_contact_survey(polymod)
} # }