contactsurveys: Download Contact Surveys for use in Infectious Disease Modelling

Provides methods for downloading contact surveys that are then used in infectious disease modelling.

See also

Author

Maintainer: Sebastian Funk sebastian.funk@lshtm.ac.uk

Authors:

Lander Willem

Hugo Gruson

Other contributors:

Maria Bekker-Nielsen Dunbar [contributor]

Carl A. B. Pearson [contributor]

Sam Clifford sj.clifford@gmail.com [contributor]

Christopher Jarvis [contributor]

Alexis Robert alexis.robert@lshtm.ac.uk [contributor]

Niel Hens [contributor]

Pietro Coletti [collector, data manager]