Download a survey from its Zenodo repository

Downloads survey data. Uses a caching mechanism via the default argument for directory.

Usage

download_survey(
  survey,
  directory = contactsurveys_dir(),
  verbose = TRUE,
  overwrite = FALSE,
  timeout = 60
)

Arguments

survey

A DOI of a survey, (see list_surveys()). If a HTML link is given, the DOI will be isolated and used.

directory

Directory of where to save survey files. The default value is to use the directory for your system using contactsurveys_dir(), which uses tools::R_user_dir(). This effectively caches the data. You can specify your own directory, or set an environment variable, CONTACTSURVEYS_HOME, see Sys.setenv() or Renviron for more detail. If this argument is set to something other than contactsurveys_dir(), a warning is issued to discourage unnecessary downloads outside the persistent cache.

verbose

Whether downloads should be echoed to output. Default TRUE.

overwrite

If files should be overwritten if they already exist. Default FALSE

timeout

A numeric value specifying timeout in seconds. Default 60 seconds.

Value

a vector of filenames, where the surveys were downloaded

See also

Examples

# \donttest{
list_surveys()
#> ℹ Downloading survey list from zenodo
#> Request failed [504]. Retrying in 1 seconds...
#> Request failed [504]. Retrying in 1.9 seconds...
#> Error in while_oai(url, args, token, as, ...): Gateway Timeout (HTTP 504).
#> ✖ Downloading survey list from zenodo [1m 34.3s]
#> 
peru_survey <- download_survey("https://doi.org/10.5281/zenodo.1095664")
#> Fetching contact survey filenames from: https://doi.org/10.5281/zenodo.1095664.
#> ✔ Successfully fetched list of published records!
#> ! No record for DOI '10.5281/zenodo.1095664'!
#> ℹ Try to get deposition by Zenodo specific record id '1095664'
#> ✔ Successfully fetched list of published records!
#> ! No record for id '1095664'!
#> ℹ Successfully fetched list of published records - page 1
#> ✔ Successfully fetched list of published records!
#> ✔ Successfully fetched published record for concept DOI '10.5281/zenodo.1095664'!
#> Downloading from https://doi.org/10.5281/zenodo.1095664.
#> ! Overwrite is 'false', aborting download of existing files
#> ℹ Download in sequential mode
#> [zen4R][INFO] ZenodoRecord - Download in sequential mode 
#> ℹ Will download 7 files from record '3874805' (doi: '10.5281/zenodo.3874805') - total size: 760.6 KiB
#> [zen4R][INFO] ZenodoRecord - Will download 7 files from record '3874805' (doi: '10.5281/zenodo.3874805') - total size: 760.6 KiB 
#> ℹ Downloading file '2015_Grijalva_Peru_sday.csv' - size: 16 KiB
#> [zen4R][INFO] Downloading file '2015_Grijalva_Peru_sday.csv' - size: 16 KiB
#> ℹ Downloading file '2015_Grijalva_Peru_participant_common.csv' - size: 11.3 KiB
#> [zen4R][INFO] Downloading file '2015_Grijalva_Peru_participant_common.csv' - size: 11.3 KiB
#> ℹ Downloading file '2015_Grijalva_Peru_hh_extra.csv' - size: 27 KiB
#> [zen4R][INFO] Downloading file '2015_Grijalva_Peru_hh_extra.csv' - size: 27 KiB
#> ℹ Downloading file '2015_Grijalva_Peru_hh_common.csv' - size: 1.5 KiB
#> [zen4R][INFO] Downloading file '2015_Grijalva_Peru_hh_common.csv' - size: 1.5 KiB
#> ℹ Downloading file '2015_Grijalva_Peru_dictionary.xls' - size: 32.5 KiB
#> [zen4R][INFO] Downloading file '2015_Grijalva_Peru_dictionary.xls' - size: 32.5 KiB
#> ℹ Downloading file '2015_Grijalva_Peru_contact_common.csv' - size: 617.3 KiB
#> [zen4R][INFO] Downloading file '2015_Grijalva_Peru_contact_common.csv' - size: 617.3 KiB
#> ℹ Downloading file '2015_Grijalva_Peru_participant_extra.csv' - size: 55 KiB
#> [zen4R][INFO] Downloading file '2015_Grijalva_Peru_participant_extra.csv' - size: 55 KiB
#> ℹ Files downloaded at '/home/runner/.local/share/R/contactsurveys'.
#> [zen4R][INFO] Files downloaded at '/home/runner/.local/share/R/contactsurveys'.
#> ℹ Verifying file integrity...
#> [zen4R][INFO] ZenodoRecord - Verifying file integrity... 
#> ℹ File '2015_Grijalva_Peru_sday.csv': integrity verified (md5sum: b43c28fa6cce8d7bd6ec6b0621aa5b02)
#> [zen4R][INFO] File '2015_Grijalva_Peru_sday.csv': integrity verified (md5sum: b43c28fa6cce8d7bd6ec6b0621aa5b02)
#> ℹ File '2015_Grijalva_Peru_participant_common.csv': integrity verified (md5sum: dada729474938c5ed593fc824f98fff9)
#> [zen4R][INFO] File '2015_Grijalva_Peru_participant_common.csv': integrity verified (md5sum: dada729474938c5ed593fc824f98fff9)
#> ℹ File '2015_Grijalva_Peru_hh_extra.csv': integrity verified (md5sum: ad68e9171a547a0634998f413c02464d)
#> [zen4R][INFO] File '2015_Grijalva_Peru_hh_extra.csv': integrity verified (md5sum: ad68e9171a547a0634998f413c02464d)
#> ℹ File '2015_Grijalva_Peru_hh_common.csv': integrity verified (md5sum: a86d50476a603f7633de8f30959b6b03)
#> [zen4R][INFO] File '2015_Grijalva_Peru_hh_common.csv': integrity verified (md5sum: a86d50476a603f7633de8f30959b6b03)
#> ℹ File '2015_Grijalva_Peru_dictionary.xls': integrity verified (md5sum: 4c43fb5e299090740663db093823fdd9)
#> [zen4R][INFO] File '2015_Grijalva_Peru_dictionary.xls': integrity verified (md5sum: 4c43fb5e299090740663db093823fdd9)
#> ℹ File '2015_Grijalva_Peru_contact_common.csv': integrity verified (md5sum: 50dcb17f306aa7c62846351c5019bdad)
#> [zen4R][INFO] File '2015_Grijalva_Peru_contact_common.csv': integrity verified (md5sum: 50dcb17f306aa7c62846351c5019bdad)
#> ℹ File '2015_Grijalva_Peru_participant_extra.csv': integrity verified (md5sum: 4d7044f954db6a83ee1b9bcc95b24cfa)
#> [zen4R][INFO] File '2015_Grijalva_Peru_participant_extra.csv': integrity verified (md5sum: 4d7044f954db6a83ee1b9bcc95b24cfa)
#> ✔ End of download
#> [zen4R][INFO] ZenodoRecord - End of download 
# }