List all surveys available for download

List all surveys available for download

Usage

list_surveys(
  directory = contactsurveys_dir(),
  overwrite = FALSE,
  verbose = TRUE
)

Arguments

directory

Directory of where to save survey files. The default value is to use the directory for your system using contactsurveys_dir(), which uses tools::R_user_dir(). This effectively caches the data. You can specify your own directory, or set an environment variable, CONTACTSURVEYS_HOME, see Sys.setenv() or Renviron for more detail. If this argument is set to something other than contactsurveys_dir(), a warning is issued to discourage unnecessary downloads outside the persistent cache.

overwrite

If files should be overwritten if they already exist. Default FALSE

verbose

Whether downloads should be echoed to output. Default TRUE.

Value

data.table with columns: date_added, title, creator, url

Examples

if (FALSE) { # \dontrun{
list_surveys()
} # }